+++ backend/src/main/java/com/example/MarketMarvels/repository/CustomerRepository.java
@@ -8,10 +8,10 @@ import java.util.List;
 public interface CustomerRepository extends JpaRepository<Customer, Long> {
     List<Customer> findByFirstName(String firstName);
 
-    List<Customer>findByid(int id);
+    List<Customer>findByid(Long id);
 
     List<Customer>findByEmail(String email);
 
     List<Customer>findByDepartment(String department);
-    
+
 }
